Changes base pkg name to fix sonarlint issue
