+++ configuration/src/main/java/com/foodgrid/configuration/ConfigurationApplication.java
@@ -1,0 +1,13 @@
+package com.foodgrid.configuration;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.config.server.EnableConfigServer;
+
+@SpringBootApplication
+@EnableConfigServer
+public class ConfigurationApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(ConfigurationApplication.class, args);
+    }
+}
