--- eureka_server/src/main/java/com/foodPuppy/eureka_server/EurekaServerApplication.java
@@ -1,15 +1,0 @@
-package com.foodPuppy.eureka_server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication
-@EnableEurekaServer
-public class EurekaServerApplication {
-    
-    public static void main(String[] args) {
-        SpringApplication.run(EurekaServerApplication.class, args);
-    }
-
-}
