+++ gateway/pom.xml
@@ -3,8 +3,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>com.foodPuppy</groupId>
-        <artifactId>FoodPuppy</artifactId>
+        <groupId>com.foodgrid</groupId>
+        <artifactId>foodgrid</artifactId>
         <version>0.0.1-SNAPSHOT</version>
     </parent>
     <artifactId>gateway</artifactId>
@@ -12,7 +12,7 @@
     <name>gateway</name>
 
     <packaging>jar</packaging>
-    <description>FoodPuppy: food delivery backend restaurant service</description>
+    <description>FoodGrid: food delivery backend restaurant service</description>
 
     <properties>
         <java.version>11</java.version>
