+++ restaurant/src/main/java/com/foodgrid/restaurant/RestaurantApplication.java
@@ -1,0 +1,25 @@
+package com.foodgrid.restaurant;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@SpringBootApplication
+@RestController
+@EnableEurekaClient
+@CrossOrigin("*")
+public class RestaurantApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(RestaurantApplication.class, args);
+    }
+
+    @GetMapping(value = {"/restaurant/", "/"})
+    public ResponseEntity<String> defaultGet() {
+        return new ResponseEntity<>("Restaurant Service ", HttpStatus.OK);
+    }
+}
