--- restaurant_service/src/main/java/com/foodPuppy/restaurant_service/RestaurantServiceApplication.java
@@ -1,25 +1,0 @@
-package com.foodPuppy.restaurant_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@SpringBootApplication
-@RestController
-@EnableEurekaClient
-@CrossOrigin("*")
-public class RestaurantServiceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(RestaurantServiceApplication.class, args);
-    }
-
-    @GetMapping(value = {"/restaurant/", "/"})
-    public ResponseEntity<String> defaultGet() {
-        return new ResponseEntity<>("Restaurant Service ", HttpStatus.OK);
-    }
-}
