--- restaurant_service/src/main/resources/bootstrap.yml
@@ -1,9 +1,0 @@
-server:
-  port: 8082
-
-spring:
-  application:
-    name: restaurant
-  cloud:
-    config:
-      uri: http://configuration:8888
