--- restaurant_service/src/test/java/com/foodPuppy/restaurant_service/RestaurantServiceApplicationTests.java
@@ -1,12 +1,0 @@
-package com.foodPuppy.restaurant_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RestaurantServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-}
