+++ user/src/main/java/com/foodgrid/user/UserApplication.java
@@ -1,0 +1,28 @@
+package com.foodgrid.user;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@SpringBootApplication
+@RestController
+@EnableEurekaClient
+@CrossOrigin("*")
+public class UserApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(UserApplication.class, args);
+    }
+
+
+    @GetMapping(value = {"/user/", "/"})
+    public ResponseEntity<String> defaultGet() {
+        return new ResponseEntity<>("User Service ", HttpStatus.OK);
+    }
+
+}
