+++ user/src/test/java/com/foodgrid/user/UserApplicationTests.java
@@ -1,0 +1,17 @@
+package com.foodgrid.user;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserApplicationTests {
+
+    @Test
+    void contextLoads() {
+        String expect = "test";
+        String actual = "test";
+        Assertions.assertEquals(expect, actual);
+    }
+
+}
