--- user_service/src/main/java/com/foodPuppy/user_service/UserServiceApplication.java
@@ -1,28 +1,0 @@
-package com.foodPuppy.user_service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@SpringBootApplication
-@RestController
-@EnableEurekaClient
-@CrossOrigin("*")
-public class UserServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(UserServiceApplication.class, args);
-    }
-
-
-    @GetMapping(value = {"/user/", "/"})
-    public ResponseEntity<String> defaultGet() {
-        return new ResponseEntity<>("User Service ", HttpStatus.OK);
-    }
-
-}
