--- user_service/src/test/java/com/foodPuppy/user_service/UserServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.foodPuppy.user_service;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
